+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
+### Fixed
 - Fixed duplicate definition with GCC 10
   [ice/framework#266](https://github.com/ice/framework/pull/266)
+
+### Changed
+- Improved stubs generation for methods which may return object or null
+  [#2092](https://github.com/phalcon/zephir/issues/2092)
 
 ## [0.12.18] - 2020-04-25
 ### Fixed